refactor non-negative adjustment
