+++ src/main/scala/edu/uci/eecs/spectralLDA/algorithm/ALSSketch.scala
@@ -11,9 +11,8 @@ import breeze.signal.{fourierTr, iFourierTr}
 import breeze.math.Complex
 import breeze.stats.median
 import edu.uci.eecs.spectralLDA.sketch.TensorSketcher
-import org.apache.spark.SparkContext
+import edu.uci.eecs.spectralLDA.utils.NonNegativeAdjustment
 
-import scalaxy.loops._
 import scala.language.postfixOps
 import scala.util.control.Breaks._
 
@@ -65,7 +64,7 @@ class ALSSketch(dimK: Int,
     val topicWordMatrix: breeze.linalg.DenseMatrix[Double] = unwhiteningMatrix * A * diag(lambda)
 
     if (nonNegativeDocumentConcentration) {
-      val topicWordMatrix_normed: breeze.linalg.DenseMatrix[Double] = simplexProj_Matrix(topicWordMatrix)
+      val topicWordMatrix_normed: DenseMatrix[Double] = NonNegativeAdjustment.simplexProj_Matrix(topicWordMatrix)
       (topicWordMatrix_normed, alpha)
     }
     else {
@@ -88,50 +87,6 @@ class ALSSketch(dimK: Int,
     // i.e T * pinv((C katri-rao dot B)^T)
     TIBC * Inverted
   }
-
-  private def simplexProj_Matrix(M :DenseMatrix[Double]): DenseMatrix[Double] ={
-    val M_onSimplex: DenseMatrix[Double] = DenseMatrix.zeros[Double](M.rows, M.cols)
-    for(i <- 0 until M.cols optimized){
-      val thisColumn = M(::,i)
-
-      val tmp1 = simplexProj(thisColumn)
-      val tmp2 = simplexProj(-thisColumn)
-      val err1:Double = breeze.linalg.norm(tmp1 - thisColumn)
-      val err2:Double = breeze.linalg.norm(tmp2 - thisColumn)
-      if(err1 > err2){
-        M_onSimplex(::,i) := tmp2
-      }
-      else{
-        M_onSimplex(::,i) := tmp1
-      }
-    }
-    M_onSimplex
-  }
-
-  private def simplexProj(V: DenseVector[Double]): DenseVector[Double]={
-    // val z:Double = 1.0
-    val len: Int = V.length
-    val U: DenseVector[Double] = DenseVector(V.copy.toArray.sortWith(_ > _))
-    val cums: DenseVector[Double] = DenseVector(AlgebraUtil.Cumsum(U.toArray).map(x => x-1))
-    val Index: DenseVector[Double] = DenseVector((1 to (len + 1)).toArray.map(x => 1.0/x.toDouble))
-    val InterVec: DenseVector[Double] = cums :* Index
-    val TobefindMax: DenseVector[Double] = U - InterVec
-    var maxIndex : Int = 0
-    // find maxIndex
-    breakable{
-      for (i <- 0 until len optimized){
-        if (TobefindMax(len - i - 1) > 0){
-          maxIndex = len - i - 1
-          break()
-        }
-      }
-    }
-    val theta: Double = InterVec(maxIndex)
-    val W: DenseVector[Double] = V.map(x => x - theta)
-    val P_norm: DenseVector[Double] = W.map(x => if (x > 0) x else 0)
-    P_norm
-  }
-
 }
 
 private[algorithm] object TensorSketchOps {
